move deploy command into package to avoid name conflict
