+++ src/commcare_cloud/commcare_cloud.py
@@ -14,7 +14,7 @@ from clint.textui import puts
 from commcare_cloud.cli_utils import print_command
 from commcare_cloud.colors import color_error
 from commcare_cloud.commands.ansible.downtime import Downtime
-from commcare_cloud.commands.deploy import Deploy
+from commcare_cloud.commands.deploy.command import Deploy
 from commcare_cloud.commands.migrations.couchdb import MigrateCouchdb
 from commcare_cloud.commands.migrations.copy_files import CopyFiles
 from commcare_cloud.commands.secrets import Secrets, MigrateSecrets
